+++ shardis-ui/config/webpack.config.factory.js
@@ -59,7 +59,8 @@ var baseWebpackConfig = {
         exclude: [
           fullPathTo('../node_modules/rxjs'),
           fullPathTo('../node_modules/@angular'),
-          fullPathTo('../node_modules/@angular2-material')
+          fullPathTo('../node_modules/@angular2-material'),
+          fullPathTo('../node_modules/ng2-webstorage')
         ]
       }
     ],