+++ shardis-ui/src/main/frontend/app/views/login/login.component.ts
@@ -30,7 +30,6 @@ export class Login implements OnInit {
         this.message = 'Bad credentials';
       }
     );
-
   }
 
   ngOnInit():any {