+++ shardis-ui/src/main/frontend/platform/browser/providers.ts
@@ -10,6 +10,8 @@ import {ROUTER_PROVIDERS} from '@angular/router';
 import {LocationStrategy} from '@angular/common/src/location/location_strategy';
 import {PathLocationStrategy} from '@angular/common/src/location/path_location_strategy';
 import {MATERIAL_PROVIDERS} from './material2';
+// ng2-webstorage
+import {NG2_WEBSTORAGE} from 'ng2-webstorage';
 
 /*
 * Application Providers/Directives/Pipes
@@ -20,6 +22,7 @@ export const APPLICATION_PROVIDERS = [
   ...FORM_PROVIDERS,
   ...HTTP_PROVIDERS,
    ...MATERIAL_PROVIDERS,
+  ...NG2_WEBSTORAGE,
   provide(LocationStrategy, { useClass: PathLocationStrategy as any }),
 ];