+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added angular2-material palette to main.scss
 * Added sidenav navigation
 * Added working login form
+* Added ng2-webstorage library
 
 ### 1.6.0 (26.05.2016)
 * Updated node to 6.2.0